Attempt to fix rootpagememleak test
