+++ sqlite/ext/comdb2/comdb2systblInt.h
@@ -42,7 +42,7 @@ int systblActivelocksInit(sqlite3 *db);
 int systblNetUserfuncsInit(sqlite3 *db);
 int systblClusterInit(sqlite3 *db);
 
-int systblNextAllowedTable(sqlite3_int64 *tabId);
+int comdb2_next_allowed_table(sqlite3_int64 *tabId);
 
 /* Simple yes/no answer for booleans */
 #define YESNO(x) ((x) ? "Y" : "N")
