+++ sqlite/ext/comdb2/keys.c
@@ -95,7 +95,7 @@ static int systblKeysOpen(sqlite3_vtab *p, sqlite3_vtab_cursor **ppCursor){
   memset(pCur, 0, sizeof(*pCur));
   *ppCursor = &pCur->base;
 
-  systblNextAllowedTable(&pCur->iRowid);
+  comdb2_next_allowed_table(&pCur->iRowid);
 
   return SQLITE_OK;
 }
@@ -133,7 +133,7 @@ static int systblKeysNext(sqlite3_vtab_cursor *cur){
     } while( pCur->iRowid < thedb->num_dbs );
   }
 
-  systblNextAllowedTable(&pCur->iRowid);
+  comdb2_next_allowed_table(&pCur->iRowid);
 
   return SQLITE_OK;
 }
