+++ sqlite/ext/comdb2/tablesizes.c
@@ -82,7 +82,7 @@ static int systblTblSizeOpen(sqlite3_vtab *p, sqlite3_vtab_cursor **ppCursor){
   memset(pCur, 0, sizeof(*pCur));
   *ppCursor = &pCur->base;
 
-  systblNextAllowedTable(&pCur->iRowid);
+  comdb2_next_allowed_table(&pCur->iRowid);
 
   return SQLITE_OK;
 }
@@ -101,7 +101,7 @@ static int systblTblSizeClose(sqlite3_vtab_cursor *cur){
 static int systblTblSizeNext(sqlite3_vtab_cursor *cur){
   systbl_tblsize_cursor *pCur = (systbl_tblsize_cursor*)cur;
   pCur->iRowid++;
-  systblNextAllowedTable(&pCur->iRowid);
+  comdb2_next_allowed_table(&pCur->iRowid);
   return SQLITE_OK;
 }
